Add files via upload
Overview
The Surf’s Up Challenge was to perform an analysis to determine the weather statistics during the summer and winter seasons in Hawaii. This analysis would be utilized to make decisions and inform business strategies about ideal months for ice cream consumption and surfing. This information would also assist in deciding on the intended location of a surf and ice cream shop.

Results
Monthly Temperature Analysis

June 
The average monthly temperature is 74.9(F), the minimum temperature is 64.0(F) and the maximum temperature was 85.0(F).

December
The temperature results for December were a minimum monthly temperature of 56.0(F), a maximum of 83.0(F), with an average of 71.0(F). 

The standard deviation for temperatures in June was smaller at (3.26) while the standard deviation for December was (3.75). This accounts for temperature fluctuations, thus proving the temperature ranges in June are lesser and thus making June warmer in comparison to December, which was colder, this was shown by the larger range in temperature. 

June was usually hotter than December as expected, but June and December had very close maximum temperatures 85.0(F) and 83.0(F) respectively. Additionally, the coldest June temperature was 8.0(F) more than the 56.0(F) of December. 

Summary 

Overall, the month of June had warmer temperatures when compared to December.

A query to gather more weather data to analyze the precipitation levels to further demonstrate the difference in the rainy days in June and December so the client may be able to capitalize on tourism.  This could be an opportunity for better business prospects for the client. 

An additional query to also determine the potential of Natural Disasters or extreme weather conditions during the months. The seasons for hurricanes, cyclones and Tsunamis are typically before December and could prove hazardous for business, even though the climate is warmer and would suggest more favourable for tourist related businesses. The possibility of different weather conditions in June in comparison to December could be a risk for the business if not considered when making these decisions. 

An analysis of the weather patterns between June and December at the specific weather stations to determine the ideal location for business and the probability of favorable weather conditions based on the data to gain more tourist.

The daily routine temperatures data for both June and December months could also be useful to determine the best hours of operations for business.